Signature-tree: cf88430dbc3c67238273dc47c82d4b435ffa00c0

+++ arch/sparc/include/asm/atomic_64.h
@@ -37,18 +37,18 @@ static inline void atomic64_set_unchecked(atomic64_unchecked_t *v, long i)
 
 extern void atomic_add(int, atomic_t *);
 extern void atomic_add_unchecked(int, atomic_unchecked_t *);
-extern void atomic64_add(int, atomic64_t *);
-extern void atomic64_add_unchecked(int, atomic64_unchecked_t *);
+extern void atomic64_add(long, atomic64_t *);
+extern void atomic64_add_unchecked(long, atomic64_unchecked_t *);
 extern void atomic_sub(int, atomic_t *);
 extern void atomic_sub_unchecked(int, atomic_unchecked_t *);
-extern void atomic64_sub(int, atomic64_t *);
-extern void atomic64_sub_unchecked(int, atomic64_unchecked_t *);
+extern void atomic64_sub(long, atomic64_t *);
+extern void atomic64_sub_unchecked(long, atomic64_unchecked_t *);
 
 extern int atomic_add_ret(int, atomic_t *);
-extern int atomic64_add_ret(int, atomic64_t *);
-extern int atomic64_add_ret_unchecked(int, atomic64_unchecked_t *);
+extern long atomic64_add_ret(long, atomic64_t *);
+extern long atomic64_add_ret_unchecked(long, atomic64_unchecked_t *);
 extern int atomic_sub_ret(int, atomic_t *);
-extern int atomic64_sub_ret(int, atomic64_t *);
+extern long atomic64_sub_ret(long, atomic64_t *);
 
 #define atomic_dec_return(v) atomic_sub_ret(1, v)
 #define atomic64_dec_return(v) atomic64_sub_ret(1, v)
@@ -140,7 +140,7 @@ static inline int atomic_add_unless(atomic_t *v, int a, int u)
 	((__typeof__((v)->counter))cmpxchg(&((v)->counter), (o), (n)))
 #define atomic64_xchg(v, new) (xchg(&((v)->counter), new))
 
-static inline int atomic64_add_unless(atomic64_t *v, long a, long u)
+static inline long atomic64_add_unless(atomic64_t *v, long a, long u)
 {
 	long c, old, new;
 	c = atomic64_read(v);
